requirements.txt
streamlit==1.38.0
deepface
Pillow==10.4.0
numpy
opencv-python-headless==4.10.0.84